# Add new `Performance/BigDecimalWithNumericArgument` cop